requirements.txt
requests>=2.31
paho-mqtt>=1.6
python-dotenv>=1.0
Flask>=2.3
asyncio-mqtt>=0.16
aiofiles>=23.2
aiodns>=3.2
aiohttp>=3.9
# Añade aquí las dependencias de tu proyecto, por ejemplo:
# fastapi==0.95.0
# uvicorn[standard]==0.21.1
